requirements-docker.txt
pandas
numpy
scikit-learn
streamlit
matplotlib
seaborn
pyyaml
altair
